+++ driver/keyboard.h
@@ -48,7 +48,7 @@ typedef enum KEY_Code_t KEY_Code_t;
 
 extern KEY_Code_t gKeyReading0;
 extern KEY_Code_t gKeyReading1;
-extern uint8_t gDebounceCounter;
+extern uint16_t gDebounceCounter;
 extern bool gWasFKeyPressed;
 
 KEY_Code_t KEYBOARD_Poll(void);
